#10242: Migrate sqrt_bw to TTNN
