+++ docs/source/ttnn/ttnn/ttnn/sqrt_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.sqrt_bw:
+
+ttnn.sqrt_bw
+##############
+
+.. autofunction:: ttnn.sqrt_bw
