+++ tt_eager/tt_dnn/op_library/backward/backward_ops.hpp
@@ -21,21 +21,6 @@ std::vector<Tensor> unary_mul_bw(
     float scalar,
     const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
 
-std::vector<std::optional<Tensor>> sqrt_bw(
-    uint8_t cq_id,
-    const Tensor& grad,
-    const Tensor& input,
-    const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG,
-    const std::vector<bool>& are_required_outputs = std::vector<bool>{true},
-    std::optional<Tensor> input_grad = std::nullopt);
-
-std::vector<std::optional<Tensor>> sqrt_bw(
-    const Tensor& grad,
-    const Tensor& input,
-    const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG,
-    const std::vector<bool>& are_required_outputs = std::vector<bool>{true},
-    std::optional<Tensor> input_grad = std::nullopt);
-
 std::vector<Tensor> binary_le_bw(
     const Tensor& grad,
     const Tensor& input,
